6-size.c

+++ 0x00-hello_world/6-size.c
@@ -1,14 +1,15 @@
 #include <stdio.h>
 /**
- * main - C program that prints the size of various types on the computer it is compiled and run on
+ * main - C program that prints the size types
+ *
  * Return: 0 (success)
  */
 int main(void)
 {
-	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(char));
-	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(int));
+	printf("Size of a char: %lu byte(s)\n", sizeof(char));
+	printf("Size of an int: %lu byte(s)\n", sizeof(int));
 	printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(long int));
-	printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(long long int));
+	printf("Size of a long long int: %lu byte(s)\n", sizeof(long long int));
 	printf("Size of a float: %lu byte(s)", (unsigned long)sizeof(float));
 	return (0);
 }
